+++ src/qt/managenamespage.cpp
@@ -5,9 +5,7 @@
 #include "nametablemodel.h"
 #include "csvmodelwriter.h"
 #include "guiutil.h"
-#include "../base58.h"
-#include "../main.h"
-#include "../hook.h"
+#include "../headers.h"
 #include "../wallet.h"
 #include "../namecoin.h"
 #include "guiconstants.h"
